컨트롤러 계층 피드백 반영 (#11)
* doc: controller 패키지 문서 갱신

> ApiSpec 인터페이스에서는 API 동작 정의, 문서화 작업 등을 담당
> 실 컨트롤러 구현체에서는 API 엔드포인트 라우팅, 서비스 레이어 브릿지 역할을 담당

* refactor(coupon): 쿠폰 컨트롤러 ApiSpec 추출

* refactor(order): 주문 컨트롤러 ApiSpec 추출

* refactor(point): 포인트 컨트롤러 ApiSpec 추출

* refactor(product): 상품 컨트롤러 ApiSpec 추출

* fix(point): GetPointResponse 오타 수정

> point로 필드명 수정 (기존: ponit)
> ~~ㅎㅎ... ㅁㅊㄱㅇㄴㅇ?~~

* fix(point): userId 헤더를 Long으로 수정

> 기존: String
> 다른 API와 일관성 유지를 위한 변경
